refactor: component 관련 폴더명 파스칼로 변경, svg 파일 폴더 변경
- toolbar 관련 svg toolbar 디렉토리로 이동
