+++ src/components/header/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import SVGIcon from '@src/utils/svg/logo.svg';
-import Button from '@src/components/common/Button';
+import SVGIcon from '@src/utils/svg/header/logo.svg';
+import Button from '@src/components/Common/Button';
 import colors from '@src/utils/colors';
 import HeaderTitle from './HeaderTitle';
 import * as StyleComponent from './style';
