+++ src/components/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Tab from '@src/components/tab';
-import Mathquill from '@src/components/mathquill';
-import ToolBar from '@src/components/toolbar';
-import Header from '@src/components/header';
-import InputSectionContainer from '@src/components/leftsection/inputSelectionSection';
-import MathSectionContainer from '@src/components/leftsection/mathSection';
-import TemplateContainer from '@src/components/leftsection/templateSection';
+import Tab from '@src/components/Tab';
+import Mathquill from '@src/components/MathQuill';
+import ToolBar from '@src/components/Toolbar';
+import Header from '@src/components/Header';
+import InputSectionContainer from '@src/components/LeftSection/InputSelectionSection';
+import MathSectionContainer from '@src/components/LeftSection/MathSection';
+import TemplateContainer from '@src/components/LeftSection/TemplateSection';
 import { rootState } from '@src/store/modules';
 import { useSelector } from 'react-redux';
