+++ src/components/toolbar/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import colors from '@src/utils/colors';
-import DRAWER from '@src/utils/svg/new-drawer.svg';
+import DRAWER from '@src/utils/svg/toolbar/drawer.svg';
 import BACKGROUND from '@src/utils/svg/background/background_icon.svg';
-import CLEAR from '@src/utils/svg/new-clear.svg';
-import BACK from '@src/utils/svg/new-back.svg';
-import FORWORD from '@src/utils/svg/new-forward.svg';
-import ToolBarButton from '@src/components/common/toolbarbutton';
+import CLEAR from '@src/utils/svg/toolbar/clear.svg';
+import BACK from '@src/utils/svg/toolbar/back.svg';
+import FORWORD from '@src/utils/svg/toolbar/forward.svg';
+import ToolBarButton from '@src/components/Common/ToolbarButton';
 import { useDispatch, useSelector } from 'react-redux';
 import { rootState } from '@src/store/modules';
 import { dropdown } from '@src/store/modules/drawerDropdown';
@@ -13,11 +13,11 @@ import {
   backgroundDropdown,
   BackgroundState,
 } from '@src/store/modules/backgroundDropdown';
-import DropDown from '@src/components/common/dropdown';
-import FontDropDown from '@src/components/toolbar/FontDropDown';
-import FontColorDropdown from '@src/components/toolbar/FontColorDropdown';
-import { RoundButton } from '@src/components/common/RoundButton/style';
-import GraphButton from '@src/components/toolbar/GraphButton';
+import DropDown from '@src/components/Common/DropDown';
+import FontDropDown from '@src/components/Toolbar/FontDropDown';
+import FontColorDropdown from '@src/components/Toolbar/FontColorDropdown';
+import { RoundButton } from '@src/components/Common/RoundButton/style';
+import GraphButton from '@src/components/Toolbar/GraphButton';
 import * as StyledComponent from './style';
 
 const ToolBar = () => {
